requirements.txt
pytubefix
sqlalchemy
niconico.py
psycopg
django
flask
